+++ cdx_core/Gemfile
@@ -15,6 +15,10 @@ gemspec
 
 gem 'responders', '~> 2.0'
 gem 'mysql2'
+gem 'jquery-rails'
+gem 'haml-rails'
+gem 'react-rails', '~> 1.3.2'
+gem 'react-rails-img'
 gem 'devise', '~> 3.5.5'
 gem 'devise_security_extension', git: 'https://github.com/phatworx/devise_security_extension'
 gem 'devise_invitable'
@@ -35,7 +39,8 @@ gem 'encryptor'
 gem 'base58'
 gem 'nuntium_api', '~> 0.21'
 gem 'config'
-gem 'decent_exposure'
+gem 'decent_exposure', '2.3.2'
+gem 'gon'
 
 group :development do
   gem 'letter_opener'
