+++ cdx_core/spec/dummy/config/initializers/session_store.rb
@@ -1,3 +1,5 @@
 # Be sure to restart your server when you modify this file.
 
 Rails.application.config.session_store :cookie_store, key: '_dummy_session'
+
+Rails.application.config.version_name = 'test'
